Player walks on terrain

+++ SanityEngine/src/player/flycam_controller.cpp
@@ -63,8 +63,8 @@ void FlycamController::update_player_transform(const float delta_time) {
     // Make sure they're on the terrain
     if(terrain) {
         const auto height = terrain->get_terrain_height(
-            TerrainSamplerParams{.latitude = static_cast<uint32_t>(player_transform.position.z),
-                                 .longitude = static_cast<uint32_t>(player_transform.position.x)});
+            TerrainSamplerParams{.latitude = player_transform.position.z,
+                                 .longitude = player_transform.position.x});
         if(player_transform.position.y < height + 1.5) {
             player_transform.position.y = height + 1.5;
         }
